requirements.txt
streamlit==1.35.0
python-binance==1.0.17
python-dotenv